\name{ExamSolution1-package}
\alias{ExamSolution1-package}
\alias{ExamSolution1}
\docType{package}
\title{
\packageTitle{ExamSolution1}
}
\description{
\packageDescription{ExamSolution1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ExamSolution1}
\packageIndices{ExamSolution1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ExamSolution1}

Maintainer: \packageMaintainer{ExamSolution1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
